Fixes + restructure
