+++ ArmaForces.ArmaServerManager/Features/Hangfire/JobService.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using ArmaForces.ArmaServerManager.Features.Hangfire.Helpers;
+using ArmaForces.ArmaServerManager.Features.Hangfire.Models;
+using ArmaForces.ArmaServerManager.Features.Hangfire.Persistence;
+using ArmaForces.ArmaServerManager.Features.Hangfire.Persistence.Models;
 using CSharpFunctionalExtensions;
 using Microsoft.Extensions.Logging;
 
@@ -8,22 +10,18 @@ namespace ArmaForces.ArmaServerManager.Features.Hangfire
 {
     internal class JobService : IJobService
     {
-        private readonly IJobScheduler _jobScheduler;
         private readonly IJobStorage _jobStorage;
         private readonly ILogger<JobService> _logger;
 
-        public JobService(
-            IJobScheduler jobScheduler,
-            IJobStorage jobStorage,
-            ILogger<JobService> logger)
+        public JobService(IJobStorage jobStorage, ILogger<JobService> logger)
         {
-            _jobScheduler = jobScheduler;
             _jobStorage = jobStorage;
             _logger = logger;
         }
 
         public Result<JobDetails> GetJobDetails(string jobId)
-            => _jobStorage.GetJobDetails(jobId);
+            => _jobStorage.GetJobDetails(jobId)
+                .Tap(x => _logger.LogTrace("Successfully retrieved details for job {JobId}", jobId));
 
         public Result<List<JobDetails>> GetQueuedJobs()
             => GetJobs(new List<JobStatus>
@@ -35,6 +33,7 @@ namespace ArmaForces.ArmaServerManager.Features.Hangfire
 
         public Result<List<JobDetails>> GetJobs(IEnumerable<JobStatus>? jobStatusEnumerable = null)
             => _jobStorage.GetQueuedJobs()
+                .Tap(x => _logger.LogTrace("Found {Count} queued jobs", x.Count))
                 .Map(x => FilterByJobStatus(x, jobStatusEnumerable))
                 .Map(x => x.ToList());
